fix(p17): D-475 widen table64 limits to u64 — accept large-max i64 table types
A table64 declares its limits as u64 (spec §5.3.5): `(table i64 0 4294967296
funcref)` and even max = 2^64-1 are valid TYPES (validation never allocates;
only instantiation would OOM). zwasm kept TableEntry.min/max as u32 (the slice-1
simplification), so readLimits' u32 LEB read overflowed and rejected these valid
modules (spec dir table.30/table.32 = ValidateFailed).

Widen TableEntry + TableInstance + the C-API table handle min/max to u64; a
legacy i32 table still must fit u32 (decoder-enforced). The wasm-c-api / JIT /
AOT surfaces stay u32 and saturate a wider i64-table bound to u32-max (the JIT
guard keeps i64 tables off the codegen path entirely). Unit tests: i64 table
with max = 2^32 parses; i32 table with >u32 limit rejects.

Closes the table.30/32 validator over-rejection. The last remaining table64
spec gap is cross-module (register) i64-table imports (table.12/34, table_grow.6/7
= UnknownImport), tracked in D-475.

+++ src/api/handles.zig
@@ -155,8 +155,10 @@ pub const Table = struct {
     instance: ?*Instance,
     table_idx: u32 = 0,
     elem_type: zir.ValType,
-    min: u32,
-    max: ?u32,
+    // u64 to carry a table64 (i64-indexed) table's limits; the C-API
+    // getters (wasm_table_type / wasm_tabletype_limits) saturate to u32.
+    min: u64,
+    max: ?u64,
     /// Cached borrowed extern view (see `Func.extern_view`).
     extern_view: ?*Extern = null,
     /// Cached borrowed `wasm_ref_t` view (`wasm_table_as_ref`, ADR-0158;
